+++ js/CHANGELOG.md
@@ -10,6 +10,7 @@
 - Support ReplyCode from `@gear-js/api` in https://github.com/gear-tech/sails/pull/893
 From now on, `response` function will throw an error if the program's reply is successful.
 - Update types in ctor generation (https://github.com/gear-tech/sails/issues/786)
+- Unpin `sails-js` peer dependencies
 
 ## 0.3.2